Build: Add -Werror and -march=armv7ve
To make sure we're emitting good code, all warnings now generate
errors to make sure we're really sticking to the standard as well
as all the warnings we switch one.

As this is now the case, the `-march` option has been changed to
armv7ve. This is basically the same as `-march=armv7-a`, except
it has support for the Virtualization Extensions. The reason
armv7-a was generating a warning was due to the fact that the
Cortex A-15 has some features that armv7-a does not necessarily
support.