Build: Change RAM size and modify linker script
ARM doesn't contain a method to detect the amount of RAM installed in
the system (as it's use is for a fixed/known device, instead of having
a number of different configurations). The linker script now reflects
this by defining the values of `__RAM_BASE_ADDRESS` and
`__RAM_TOP_ADDRESS`.

Some other, nice symbols have also been defined to make life a bit
easier inside the kernel.