Meta: Initial Commit
Initial Commit of previous System-V ARM clone. This time we're
targeting a different platform that has the RAM starting from address
zero instead of at 2GiB. This will make paging a lot nicer in
the long run.

We also don't have to deal with the stupid GIC anymore, as this
specific ARM board comes with a Vectored Interrupt Controller.

The documentation for the baseboard (including the memory map)
is provided in `docs/`